Remove iron from code
