+++ exonum/Cargo.toml
@@ -32,14 +32,6 @@ clap = "2.31.2"
 term = "=0.5.1"
 colored = "1.6.0"
 env_logger = "=0.5.10"
-hyper = "=0.10.13"
-cookie = "=0.10.1"
-router = "=0.6.0"
-iron = "=0.6.0"
-iron-cors = "=0.7.0"
-headers = "=0.1.0"
-params = "=0.8.0"
-mount = "=0.4.0"
 atty = "=0.2.10"
 bytes = "=0.4.8"
 futures = "=0.1.21"
