+++ exonum/src/api/error.rs
