+++ exonum/src/api/state.rs
