+++ exonum/src/api/with.rs
