+++ exonum/src/blockchain/service.rs
@@ -15,22 +15,20 @@
 //! This module defines the Exonum services interfaces. Like smart contracts in some other
 //! blockchain platforms, Exonum services encapsulate business logic of the blockchain application.
 
-use iron::Handler;
 use serde_json::Value;
 
 use std::{collections::{HashMap, HashSet},
-          fmt,
           net::SocketAddr,
           sync::{Arc, RwLock}};
 
 use super::transaction::Transaction;
-use api_ng::ServiceApiBuilder;
-use blockchain::{Blockchain, ConsensusConfig, Schema, StoredConfiguration, ValidatorKeys};
+use api::ServiceApiBuilder;
+use blockchain::{ConsensusConfig, Schema, StoredConfiguration, ValidatorKeys};
 use crypto::{Hash, PublicKey, SecretKey};
 use encoding::Error as MessageError;
 use helpers::{Height, Milliseconds, ValidatorId};
 use messages::RawTransaction;
-use node::{ApiSender, Node, State, TransactionSend};
+use node::{ApiSender, State, TransactionSend};
 use storage::{Fork, Snapshot};
 
 /// A trait that describes the business logic of a certain service.
@@ -194,24 +192,6 @@ pub trait Service: Send + Sync + 'static {
     ///
     /// *Try not to perform long operations in this handler*.
     fn after_commit(&self, context: &ServiceContext) {}
-
-    /// Returns an API handler for public requests. The handler is mounted on
-    /// the `/api/services/{service_name}` path at [the public listen address][pub-addr]
-    /// of all full nodes in the blockchain network.
-    ///
-    /// [pub-addr]: ../node/struct.NodeApiConfig.html#structfield.public_api_address
-    fn public_api_handler(&self, context: &ApiContext) -> Option<Box<Handler>> {
-        None
-    }
-
-    /// Returns an API handler for private requests. The handler is mounted on
-    /// the `/api/services/{service_name}` path at [the private listen address][private-addr]
-    /// of all full nodes in the blockchain network.
-    ///
-    /// [private-addr]: ../node/struct.NodeApiConfig.html#structfield.private_api_address
-    fn private_api_handler(&self, context: &ApiContext) -> Option<Box<Handler>> {
-        None
-    }
 
     /// Extends API by handlers of this service. The request handlers are mounted on the
     /// the `/api/services/{service_name}` path at the listen address of all
@@ -490,73 +470,6 @@ impl SharedNodeState {
             .expect("Expected write lock")
             .reconnects_timeout
             .remove(addr)
-    }
-}
-
-/// Provides the current node state to API handlers.
-pub struct ApiContext {
-    blockchain: Blockchain,
-    node_channel: ApiSender,
-    public_key: PublicKey,
-    secret_key: SecretKey,
-}
-
-/// Provides the current node state to API handlers.
-impl ApiContext {
-    /// Constructs context for the given `Node`.
-    pub fn new(node: &Node) -> ApiContext {
-        let handler = node.handler();
-        ApiContext {
-            blockchain: handler.blockchain.clone(),
-            node_channel: node.channel(),
-            public_key: *node.state().service_public_key(),
-            secret_key: node.state().service_secret_key().clone(),
-        }
-    }
-
-    /// Constructs context from raw parts.
-    pub fn from_parts(
-        blockchain: &Blockchain,
-        node_channel: ApiSender,
-        public_key: &PublicKey,
-        secret_key: &SecretKey,
-    ) -> ApiContext {
-        ApiContext {
-            blockchain: blockchain.clone(),
-            node_channel,
-            public_key: *public_key,
-            secret_key: secret_key.clone(),
-        }
-    }
-
-    /// Returns a reference to the blockchain of this node.
-    pub fn blockchain(&self) -> &Blockchain {
-        &self.blockchain
-    }
-
-    /// Returns a reference to the transaction sender.
-    pub fn node_channel(&self) -> &ApiSender {
-        &self.node_channel
-    }
-
-    /// Returns the public key of the current node.
-    pub fn public_key(&self) -> &PublicKey {
-        &self.public_key
-    }
-
-    /// Returns the secret key of the current node.
-    pub fn secret_key(&self) -> &SecretKey {
-        &self.secret_key
-    }
-}
-
-impl ::std::fmt::Debug for ApiContext {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "ApiContext(blockchain: {:?}, public_key: {:?})",
-            self.blockchain, self.public_key
-        )
     }
 }
