+++ testkit/Cargo.toml
@@ -24,11 +24,7 @@ bodyparser = "=0.8.0"
 exonum = { version = "0.8.0", path = "../exonum" }
 failure = "=0.1.1"
 futures = "=0.1.21"
-iron = "=0.6.0"
-iron-test = "=0.6.0"
 log = "=0.4.1"
-mount = "=0.4.0"
-router = "=0.6.0"
 serde = "1.0.10"
 serde_derive = "1.0.10"
 serde_json = "1.0.2"
