+++ testkit/src/api.rs
@@ -12,31 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use exonum::api_ng::ApiAccess;
+//! API encapsulation for the testkit.
 
-use bodyparser;
-use exonum::{api::ApiError,
-             blockchain::{SharedNodeState, Transaction},
-             node::{create_private_api_handler, create_public_api_handler, ApiSender,
-                    TransactionSend}};
-use iron::{headers::{ContentType, Headers},
-           status::{self, StatusClass},
-           Chain,
-           Handler,
-           IronError,
-           IronResult,
-           Plugin,
-           Request,
-           Response};
-use iron_test::{request, response};
-use log::Level;
-use mount::{Mount, OriginalUrl};
-use serde::{Deserialize, Serialize};
-use serde_json::{self, Value as JsonValue};
+pub use exonum::api::ApiAccess;
+
+use actix_web::http::Method;
+use actix_web::{test::TestServer, App, HttpMessage};
+use failure;
+use serde_urlencoded;
+use serde_json;
 
 use std::fmt;
 
-use super::TestKit;
+use exonum::{api::{ApiAggregator, ServiceApiState},
+             blockchain::{SharedNodeState, Transaction},
+             encoding::serialize::reexport::{DeserializeOwned, Serialize},
+             node::{ApiSender, TransactionSend}};
+
+use TestKit;
 
 /// Kind of public or private REST API of an Exonum node.
 ///
@@ -63,21 +56,13 @@ impl ::fmt::Display for ApiKind {
     }
 }
 
-impl ApiKind {
-    fn into_prefix(self) -> String {
-        match self {
-            ApiKind::System => "api/system".to_string(),
-            ApiKind::Explorer => "api/explorer".to_string(),
-            ApiKind::Service(name) => format!("api/services/{}", name),
-        }
-    }
-}
+/// TODO
+pub type Error = failure::Error;
 
 /// API encapsulation for the testkit. Allows to execute and synchronously retrieve results
 /// for REST-ful endpoints of services.
 pub struct TestKitApi {
-    public_handler: Chain,
-    private_handler: Chain,
+    test_server: TestServer,
     api_sender: ApiSender,
 }
 
@@ -89,64 +74,14 @@ impl fmt::Debug for TestKitApi {
 
 impl TestKitApi {
     /// Creates a new instance of API.
-    pub(crate) fn new(testkit: &TestKit) -> Self {
-        let blockchain = &testkit.blockchain;
-        let api_state = SharedNodeState::new(10_000);
+    pub fn new(testkit: &TestKit) -> Self {
+        let aggregator =
+            ApiAggregator::new(testkit.blockchain().clone(), SharedNodeState::new(10_000));
 
         TestKitApi {
-            public_handler: {
-                let mut handler = create_public_api_handler(
-                    blockchain.clone(),
-                    api_state.clone(),
-                    &testkit.api_config,
-                );
-                handler.link_after(|req: &mut Request, resp| {
-                    log_request(&ApiAccess::Public, req, resp)
-                });
-                handler
-            },
-
-            private_handler: {
-                let mut handler = create_private_api_handler(
-                    blockchain.clone(),
-                    api_state,
-                    testkit.api_sender.clone(),
-                    &testkit.api_config,
-                );
-                handler.link_after(|req: &mut Request, resp| {
-                    log_request(&ApiAccess::Private, req, resp)
-                });
-                handler
-            },
-
+            test_server: create_test_server(aggregator),
             api_sender: testkit.api_sender.clone(),
         }
-    }
-
-    /// Returns the mounting point for public APIs. Useful for intricate testing not covered
-    /// by `get*` and `post*` functions.
-    pub fn public_handler(&self) -> &Chain {
-        &self.public_handler
-    }
-
-    /// Returns the mounting point for private APIs. Useful for intricate testing not covered
-    /// by `get*` and `post*` functions.
-    pub fn private_handler(&self) -> &Chain {
-        &self.private_handler
-    }
-
-    pub(crate) fn into_handlers<H: Handler>(self, testkit_handler: H) -> (Chain, Chain) {
-        let (public_handler, private_handler) = (self.public_handler, self.private_handler);
-
-        let mut testkit_handler = Chain::new(testkit_handler);
-        testkit_handler
-            .link_after(|req: &mut Request, resp| log_request(&ApiAccess::Private, req, resp));
-
-        let mut private_mount = Mount::new();
-        private_mount.mount("api/testkit", testkit_handler);
-        private_mount.mount("", private_handler);
-
-        (public_handler, Chain::new(private_mount))
     }
 
     /// Sends a transaction to the node via `ApiSender`.
@@ -159,323 +94,137 @@ impl TestKitApi {
             .expect("Cannot send transaction");
     }
 
-    fn get_internal<H, D>(handler: &H, endpoint: &str, expect_error: bool) -> D
-    where
-        H: Handler,
-        for<'de> D: Deserialize<'de>,
-    {
-        let status_class = if expect_error {
-            StatusClass::ClientError
-        } else {
-            StatusClass::Success
-        };
-
-        let url = format!("http://localhost:3000/{}", endpoint);
-        let resp = request::get(&url, Headers::new(), handler);
-        let resp = if expect_error {
-            // Support either "normal" or erroneous responses.
-            // For example, `Api.not_found_response()` returns the response as `Ok(..)`.
-            match resp {
-                Ok(resp) => resp,
-                Err(IronError { response, .. }) => response,
-            }
-        } else {
-            resp.expect("Got unexpected `Err(..)` response")
-        };
-
-        if let Some(ref status) = resp.status {
-            if status.class() != status_class {
-                panic!("Unexpected response status: {:?}", status);
-            }
-        } else {
-            panic!("Response status not set");
-        }
-
-        let resp = response::extract_body_to_string(resp);
-        serde_json::from_str(&resp).unwrap()
+    /// TODO
+    pub fn public(&mut self, kind: ApiKind) -> RequestBuilder {
+        RequestBuilder::new(&mut self.test_server, ApiAccess::Public, kind)
     }
 
-    /// Gets information from a public endpoint of the node.
-    ///
-    /// # Panics
-    ///
-    /// - Panics if an error occurs during request processing (e.g., the requested endpoint is
-    ///  unknown), or if the response has a non-20x response status.
-    pub fn get<D>(&self, kind: ApiKind, endpoint: &str) -> D
-    where
-        for<'de> D: Deserialize<'de>,
-    {
-        TestKitApi::get_internal(
-            &self.public_handler,
-            &format!("{}/{}", kind.into_prefix(), endpoint),
-            false,
-        )
-    }
-
-    /// Gets information from a private endpoint of the node.
-    ///
-    /// # Panics
-    ///
-    /// - Panics if an error occurs during request processing (e.g., the requested endpoint is
-    ///  unknown), or if the response has a non-20x response status.
-    pub fn get_private<D>(&self, kind: ApiKind, endpoint: &str) -> D
-    where
-        for<'de> D: Deserialize<'de>,
-    {
-        TestKitApi::get_internal(
-            &self.private_handler,
-            &format!("{}/{}", kind.into_prefix(), endpoint),
-            false,
-        )
-    }
-
-    /// Gets an error from a public endpoint of the node.
-    ///
-    /// # Panics
-    ///
-    /// - Panics if the response has a non-error response status.
-    pub fn get_err(&self, kind: ApiKind, endpoint: &str) -> ApiError {
-        let url = format!("http://localhost:3000/{}/{}", kind.into_prefix(), endpoint);
-        let response = match request::get(&url, Headers::new(), &self.public_handler) {
-            Ok(response) | Err(IronError { response, .. }) => response,
-        };
-        TestKitApi::response_to_api_error(response)
-    }
-
-    fn post_internal<H, T, D>(handler: &H, endpoint: &str, data: &T) -> D
-    where
-        H: Handler,
-        T: Serialize,
-        for<'de> D: Deserialize<'de>,
-    {
-        let url = format!("http://localhost:3000/{}", endpoint);
-        let body = serde_json::to_string(&data).expect("Cannot serialize data to JSON");
-        let resp = request::post(
-            &url,
-            {
-                let mut headers = Headers::new();
-                headers.set(ContentType::json());
-                headers
-            },
-            &body,
-            handler,
-        ).expect("Cannot send data");
-
-        let resp = response::extract_body_to_string(resp);
-        serde_json::from_str(&resp).expect("Cannot parse result")
-    }
-
-    /// Posts a transaction to the service using the public API. The returned value is the result
-    /// of synchronous transaction processing, which includes running the API shim
-    /// and `Transaction.verify()`. `Transaction.execute()` is not run until the transaction
-    /// gets to a block via one of `create_block*()` methods.
-    ///
-    /// # Panics
-    ///
-    /// - Panics if an error occurs during request processing (e.g., the requested endpoint is
-    ///  unknown).
-    pub fn post<T, D>(&self, kind: ApiKind, endpoint: &str, transaction: &T) -> D
-    where
-        T: Serialize,
-        for<'de> D: Deserialize<'de>,
-    {
-        TestKitApi::post_internal(
-            &self.public_handler,
-            &format!("{}/{}", kind.into_prefix(), endpoint),
-            transaction,
-        )
-    }
-
-    /// Posts a transaction to the service using the private API. The returned value is the result
-    /// of synchronous transaction processing, which includes running the API shim
-    /// and `Transaction.verify()`. `Transaction.execute()` is not run until the transaction
-    /// gets to a block via one of `create_block*()` methods.
-    ///
-    /// # Panics
-    ///
-    /// - Panics if an error occurs during request processing (e.g., the requested endpoint is
-    ///  unknown).
-    pub fn post_private<T, D>(&self, kind: ApiKind, endpoint: &str, transaction: &T) -> D
-    where
-        T: Serialize,
-        for<'de> D: Deserialize<'de>,
-    {
-        TestKitApi::post_internal(
-            &self.private_handler,
-            &format!("{}/{}", kind.into_prefix(), endpoint),
-            transaction,
-        )
-    }
-
-    /// Converts iron Response to ApiError.
-    ///
-    /// # Panics
-    ///
-    /// - Panics if the response has a non-error response status.
-    fn response_to_api_error(response: Response) -> ApiError {
-        fn extract_description(body: &str) -> Option<String> {
-            match serde_json::from_str::<JsonValue>(body).ok()? {
-                JsonValue::Object(ref object) if object.contains_key("description") => {
-                    Some(object["description"].as_str()?.to_owned())
-                }
-                JsonValue::String(string) => Some(string),
-                _ => None,
-            }
-        }
-
-        fn error(response: Response) -> String {
-            let body = response::extract_body_to_string(response);
-            extract_description(&body).unwrap_or(body)
-        }
-
-        let status = response.status.expect("Status header is not set");
-
-        match status {
-            status::Forbidden => ApiError::Unauthorized,
-            status::BadRequest => ApiError::BadRequest(error(response)),
-            status::NotFound => ApiError::NotFound(error(response)),
-            s if s.is_server_error() => ApiError::InternalError(error(response).into()),
-            s => panic!("Received non-error response status: {}", s.to_u16()),
-        }
+    /// TODO
+    pub fn private(&mut self, kind: ApiKind) -> RequestBuilder {
+        RequestBuilder::new(&mut self.test_server, ApiAccess::Private, kind)
     }
 }
 
-// Logging middleware for `TestKitApi`.
-fn log_request(
-    access: &ApiAccess,
-    request: &mut Request,
-    mut response: Response,
-) -> IronResult<Response> {
-    use iron::headers::ContentLength;
-    use iron::method::Method;
-    use iron::response::WriteBody;
+/// TODO
+pub struct RequestBuilder<'a, Q = ()> {
+    test_server: &'a mut TestServer,
+    access: ApiAccess,
+    kind: ApiKind,
+    query: Option<Q>,
+}
 
-    fn has_body(method: &Method) -> bool {
-        match *method {
-            Method::Get | Method::Head | Method::Delete | Method::Trace => false,
-            _ => true,
+impl<'a, Q> fmt::Debug for RequestBuilder<'a, Q>
+where
+    Q: fmt::Debug + Serialize,
+{
+    fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
+        f.debug_struct("RequestBuilder")
+            .field("access", &self.access)
+            .field("kind", &self.kind)
+            .field("query", &self.query)
+            .finish()
+    }
+}
+
+impl<'a, Q> RequestBuilder<'a, Q>
+where
+    Q: Serialize,
+{
+    /// TODO
+    pub fn new(test_server: &'a mut TestServer, access: ApiAccess, kind: ApiKind) -> Self {
+        RequestBuilder {
+            test_server,
+            access,
+            kind,
+            query: None,
         }
     }
 
-    if !log_enabled!(Level::Trace) {
-        // Avoid expensive string allocations.
-        return Ok(response);
+    ///TODO
+    pub fn query<T>(self, query: T) -> RequestBuilder<'a, T> {
+        RequestBuilder {
+            test_server: self.test_server,
+            access: self.access,
+            kind: self.kind,
+            query: Some(query),
+        }
     }
 
-    let mut url = request
-        .extensions
-        .get::<OriginalUrl>()
-        .unwrap_or(&request.url)
-        .path()
-        .join("/");
-    if let Some(query) = request.url.query() {
-        url += "?";
-        url += query;
-    }
+    /// TODO
+    pub fn get<R>(&mut self, endpoint: &str) -> Result<R, Error>
+    where
+        R: DeserializeOwned + 'static,
+    {
+        let kind = self.kind;
+        let access = self.access;
 
-    let req_body: Option<String> = if has_body(&request.method) {
-        request
-            .get::<bodyparser::Raw>()
-            .expect("cannot read request body")
-    } else {
-        None
-    };
-
-    let response_body: Option<String> = {
-        let len: u64 = response
-            .headers
-            .get::<ContentLength>()
-            .map(|len| len.0)
+        let params = self.query
+            .as_ref()
+            .map(|query| serde_urlencoded::to_string(query).expect("Unable to serialize query."))
             .unwrap_or_default();
+        let path = format!("{}/{}/{}{}", access, kind, endpoint, params);
 
-        response.body.take().map(|mut body| {
-            let mut buffer = Vec::with_capacity(len as usize);
-            body.write_body(&mut buffer)
-                .expect("cannot write response body to buffer");
-            String::from_utf8(buffer).expect("cannot decode response body")
-        })
-    };
+        trace!("GET: {}", self.test_server.url(&path));
 
-    trace!(
-        "{method} ({access}) /{url}{req_body_tag}{req_body}\n\
-         Response: {resp_status}\n{resp_body}\n",
-        method = request.method,
-        access = access,
-        url = url,
-        req_body_tag = if req_body.is_some() { "\nBody: " } else { "" },
-        req_body = req_body.unwrap_or_default(),
-        resp_status = response
-            .status
-            .as_ref()
-            .map(ToString::to_string)
-            .unwrap_or_else(|| "(no status)".to_string()),
-        resp_body = response_body
-            .as_ref()
-            .map(String::as_ref)
-            .unwrap_or("(no body)")
-    );
+        let request = self.test_server
+            .client(Method::GET, &path)
+            .finish()
+            .expect("WTF")
+            .send();
 
-    // Return the body to the response
-    response.body = response_body.map(|body| Box::new(body) as Box<WriteBody>);
-    Ok(response)
+        let response = self.test_server.execute(request)?;
+
+        trace!("Response: {:?}", response);
+
+        self.test_server
+            .execute(response.json())
+            .map_err(From::from)
+    }
+
+    /// TODO
+    pub fn post<R>(&mut self, endpoint: &str) -> Result<R, Error>
+    where
+        R: DeserializeOwned + 'static,
+    {
+        let kind = self.kind;
+        let access = self.access;
+        let path = format!("{}/{}/{}", access, kind, endpoint);
+
+        trace!("POST: {}", self.test_server.url(&path));
+
+        let mut request = self.test_server.client(Method::POST, &path);
+        let request = if let Some(ref query) = self.query.as_ref() {
+            trace!("Body: {}", serde_json::to_string_pretty(query).unwrap());
+            request.json(query)
+        } else {
+            request.json(&())
+        }.expect("WTF")
+            .send();
+
+        let response = self.test_server.execute(request)?;
+
+        trace!("Response: {:?}", response);
+
+        self.test_server
+            .execute(response.json())
+            .map_err(From::from)
+    }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
+/// Creates test server.
+fn create_test_server(aggregator: ApiAggregator) -> TestServer {
+    let server = TestServer::with_factory(move || {
+        let state = ServiceApiState::new(aggregator.blockchain());
+        App::with_state(state.clone())
+            .scope("public/api", |scope| {
+                aggregator.extend_api(ApiAccess::Public, scope)
+            })
+            .scope("private/api", |scope| {
+                aggregator.extend_api(ApiAccess::Private, scope)
+            })
+    });
 
-    #[test]
-    fn test_get_err_non_json() {
-        let response = Response::with((status::NotFound, "Not found"));
-        assert_matches!(
-            TestKitApi::response_to_api_error(response),
-            ApiError::NotFound(ref body) if body == "Not found"
-        );
-    }
+    info!("Test server created on {}", server.addr());
 
-    #[test]
-    fn test_get_err_json_string() {
-        let response = Response::with((status::NotFound, "\"Wallet not found\""));
-        assert_matches!(
-            TestKitApi::response_to_api_error(response),
-            ApiError::NotFound(ref body) if body == "Wallet not found"
-        );
-    }
-
-    #[test]
-    fn test_get_err_json_object_with_description() {
-        let response_body = r#"{ "debug": "Some debug info", "description": "Some description" }"#;
-        let response = Response::with((status::BadRequest, response_body));
-        assert_matches!(
-            TestKitApi::response_to_api_error(response),
-            ApiError::BadRequest(ref body) if body == "Some description"
-        );
-    }
-
-    #[test]
-    fn test_get_err_json_object_without_description() {
-        let response_body = r#"{ "type": "unknown" }"#;
-        let response = Response::with((status::BadRequest, response_body));
-        assert_matches!(
-            TestKitApi::response_to_api_error(response),
-            ApiError::BadRequest(ref body) if body == response_body
-        );
-    }
-
-    #[test]
-    fn test_get_err_other_json() {
-        let response_body = r#"[1, 2, 3]"#;
-        let response = Response::with((status::BadRequest, response_body));
-        assert_matches!(
-            TestKitApi::response_to_api_error(response),
-            ApiError::BadRequest(ref body) if body == response_body
-        );
-    }
-
-    #[test]
-    #[should_panic(expected = "Received non-error response status")]
-    fn test_get_err_non_error_status() {
-        let response = Response::with(status::Ok);
-        TestKitApi::response_to_api_error(response);
-    }
+    server
 }
